harden(media): validate the upload boundary on both surfaces
Neither `/api/media/*` nor `/v1/media/*` checked the *shape* of what a
client sent, only its presence:

- `fileSize` was read straight into the presigned ContentLength and into
  the length comparison against the stored object. A non-number passes both
  quietly — `[] > max` is false and `contentLength - []` is not NaN — so the
  two checks that bound an upload could both be no-ops.
- `storageFilename` became the storage key verbatim. Presign only ever mints
  `<uuid>.<ext>`, so require exactly that: a name containing `..` still
  matches the `uploads/<user>/` prefix check, and object storage keeps such a
  key as written while a CDN in front of it may resolve it.

Neither is a proven exploit — the signed ContentLength still binds the
upload, and only presign can create an object — but these are the two
values the whole upload contract rests on and they were unvalidated.

The filename rule lives in `media-upload-policy.ts` so the session and API
key surfaces cannot drift apart.

Verified against a running API with R2 configured: string, array, negative,
fractional and oversized sizes are 400, a disallowed content type is 400, a
legitimate presign still returns a signed URL, and confirm refuses both a
traversal filename and another user's prefix.

+++ backend/server/src/routes/v1/media.ts
@@ -7,7 +7,7 @@ import { v1ConfirmMedia, v1GetMedia, v1PresignMedia } from "../../services/v1-me
 const presignSchema = z.object({
   filename: z.string().min(1),
   content_type: z.string().min(1),
-  size_bytes: z.number().positive(),
+  size_bytes: z.number().int().positive(),
 });
 
 const confirmSchema = z.object({
@@ -15,7 +15,7 @@ const confirmSchema = z.object({
   storage_filename: z.string().min(1),
   original_filename: z.string().min(1),
   content_type: z.string().min(1),
-  size_bytes: z.number().positive(),
+  size_bytes: z.number().int().positive(),
 });
 
 export async function registerMediaRoutes(app: FastifyInstance) {
